requirements.txt
streamlit
folium
streamlit-folium
requests
pandas
python-dotenv
# konlpy # Optional for Phase 2
